Moved to protcol definitions for operations to allow Mock and production types without dependencies
